requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==2.28.0
reportlab==4.2.0
matplotlib==3.9.0
numpy==1.26.4
python-dotenv==1.0.1